Add proxmox hosts and add SSH password
Seems to only use password as a fallback, so this means it can use it if
it is set initially (without keys), then use keys going forward. Might
end up setting this to a "temp" password at some point that's only ever
used on "fresh" hosts. But for now, we can let it match.