Fix /etc/hosts FQDN
So yeah, doesn't matter what DHCP is doing, really, it was getting the
FQDN as just the hostname by default. Since we are using FQDNs for the
inventory names, we can kinda cheat to make sure we use that if the FQDN
is the same as the hostname. Might be a better option, but for now...
this is good enough.