Try this trick for the IPA web VIP
Technically, this supposedly works "by accident", but seems this
literally does what I want, "set the named header to the destination
hostname before sending the request", but their documentation suggests
that this is a "side-effect" that it works for the Host header... But
seems this is literally what it is supposed to do? Whatever, if it works
for now, good enough.